Quadratization

PseudoBooleanOptimization.quadratize — Function

quadratize(aux, f::PBF{V, T}, ::Quadratization{Q}) where {V,T,Q}

Quadratizes a given PBF, i.e., applies a mapping $\mathcal{Q} : \mathscr{F}^{k} \to \mathscr{F}^{2}$, where $\mathcal{Q}$ is the quadratization method.

Auxiliary variables

The aux function is expected to produce auxiliary variables with the following signatures:

aux()::V where {V}

Creates and returns a single variable.

aux(n::Integer)::Vector{V} where {V}

Creates and returns a vector with $n$ variables.

quadratize(f::PBF{V, T}, ::Quadratization{Q}) where {V,T,Q}

When aux is not specified, uses auxgen to generate variables.

PseudoBooleanOptimization.quadratize! — Function

quadratize!(aux, f::PBF{V, T}, ::Quadratization{Q}) where {V,T,Q}
quadratize!(f::PBF{V, T}, ::Quadratization{Q}) where {V,T,Q}

In-place version of quadratize.

PseudoBooleanOptimization.auxgen — Function

auxgen(::AbstractFunction{V,T}; name::AbstractString = "aux") where {V<:AbstractString,T}

Creates a function that, when called multiple times, returns the strings "aux_1", "aux_2", ... and so on.

auxgen(::AbstractFunction{Symbol,T}; name::Symbol = :aux) where {T}

Creates a function that, when called multiple times, returns the symbols :aux_1, :aux_2, ... and so on.

auxgen(::AbstractFunction{V,T}; start::V = V(0), step::V = V(-1)) where {V<:Integer,T}

Creates a function that, when called multiple times, returns the integers $-1$, $-2$, ... and so on.

Methods

PseudoBooleanOptimization.NTR_KZFD — Type

Quadratization{NTR_KZFD}(stable::Bool = false)

Negative term reduction NTR-KZFD (Kolmogorov & Zabih, 2004; Freedman & Drineas, 2005)

Let $f(\mathbf{x}) = -x_{1} x_{2} \dots x_{k}$.

\[\mathcal{Q}\left\lbrace{f}\right\rbrace(\mathbf{x}; z) = (k - 1) z - \sum_{i = 1}^{k} x_{i} z\]

where $\mathbf{x} \in \mathbb{B}^k$

Info

Introduces a new variable $z$ and no non-submodular terms.

PseudoBooleanOptimization.PTR_BG — Type

Quadratization{PTR_BG}(stable::Bool = false)

Positive term reduction PTR-BG (Boros & Gruber, 2014)

Let $f(\mathbf{x}) = x_{1} x_{2} \dots x_{k}$.

\[\mathcal{Q}\left\lbrace{f}\right\rbrace(\mathbf{x}; \mathbf{z}) = \left[{ \sum_{i = 1}^{k-2} z_{i} \left({ k - i - 1 + x_{i} - \sum_{j = i+1}^{k} x_{j} }\right) }\right] + x_{k-1} x_{k}\]

where $\mathbf{x} \in \mathbb{B}^k$ and $\mathbf{z} \in \mathbb{B}^{k-2}$

Info

Introduces $k - 2$ new variables $z_{1}, \dots, z_{k-2}$ and $k - 1$ non-submodular terms.

PseudoBooleanOptimization.DEFAULT — Type

Quadratization{DEFAULT}(stable::Bool = false)

Employs other methods, specifically NTR_KZFD and PTR_BG.